Rename kbn-monaco to osd-monaco (#37)
