+++ packages/osd-monaco/scripts/build.js
@@ -20,7 +20,7 @@
 const path = require('path');
 const del = require('del');
 const supportsColor = require('supports-color');
-const { run } = require('@kbn/dev-utils');
+const { run } = require('@osd/dev-utils');
 
 const TARGET_BUILD_DIR = path.resolve(__dirname, '../target');
 const ROOT_DIR = path.resolve(__dirname, '../');
