+++ packages/osd-monaco/src/monaco.ts
@@ -17,17 +17,20 @@
  * under the License.
  */
 
-/* eslint-disable-next-line @kbn/eslint/module_migration */
-import * as monaco from 'monaco-editor/esm/vs/editor/editor.api';
-import * as xJson from './xjson';
-import * as esql from './esql';
-import * as painless from './painless';
+/* eslint-disable @osd/eslint/module_migration */
 
-export const registerLexerRules = (m: typeof monaco) => {
-  m.languages.register({ id: xJson.ID });
-  m.languages.setMonarchTokensProvider(xJson.ID, xJson.lexerRules);
-  m.languages.register({ id: painless.ID });
-  m.languages.setMonarchTokensProvider(painless.ID, painless.lexerRules);
-  m.languages.register({ id: esql.ID });
-  m.languages.setMonarchTokensProvider(esql.ID, esql.lexerRules);
-};
+import * as monaco from 'monaco-editor/esm/vs/editor/editor.api';
+
+import 'monaco-editor/esm/vs/base/common/worker/simpleWorker';
+import 'monaco-editor/esm/vs/base/worker/defaultWorkerFactory';
+
+import 'monaco-editor/esm/vs/editor/browser/controller/coreCommands.js';
+import 'monaco-editor/esm/vs/editor/browser/widget/codeEditorWidget.js';
+
+import 'monaco-editor/esm/vs/editor/contrib/wordOperations/wordOperations.js'; // Needed for word-wise char navigation
+
+import 'monaco-editor/esm/vs/editor/contrib/suggest/suggestController.js'; // Needed for suggestions
+import 'monaco-editor/esm/vs/editor/contrib/hover/hover.js'; // Needed for hover
+import 'monaco-editor/esm/vs/editor/contrib/parameterHints/parameterHints.js'; // Needed for signature
+
+export { monaco };
