+++ packages/osd-monaco/src/xjson/README.md
