+++ packages/osd-monaco/src/xjson/lexer_rules/esql.ts
