+++ packages/osd-monaco/src/xjson/lexer_rules/painless.ts
