+++ packages/osd-monaco/src/xjson/lexer_rules/shared.ts
