+++ packages/osd-monaco/src/xjson/lexer_rules/xjson.ts
