+++ packages/osd-monaco/tsconfig.json
